Extract SQLite helper methods from task worker
